# many speed optimizations in Adafruit_ILI9341 lib (3x times faster)