--- a/libraries/Adafruit_ILI9341/Adafruit_ILI9341.h
+++ b/libraries/Adafruit_ILI9341/Adafruit_ILI9341.h
@@ -161,16 +161,20 @@ class Adafruit_ILI9341 : public Adafruit_GFX {
 
   uint8_t  spiread(void);
 
+
 #ifdef ESP8266
   inline void spiwrite(uint8_t data);
   inline void spiwrite16(uint16_t data);
   inline void spiwriteBytes(uint8_t * data, uint8_t size);
   inline void spiwritePattern(uint8_t * data, uint8_t size, uint32_t repeat);
+
+  inline void setAddrWindow_(uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y1);
 #else
   void spiwrite(uint8_t);
   void spiwrite16(uint16_t data);
   void spiwriteBytes(uint8_t * data, uint8_t size);
   void spiwritePattern(uint8_t * data, uint8_t size, uint8_t repeat);
+  void setAddrWindow_(uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y1);
 #endif
 
   inline void spiCsHigh(void);