… hardware/esp8266com/esp8266/platform.txt
… libraries/SD/src/SD.cpp
… libraries/SD/src/SD.h
… libraries/SD/src/utility/Sd2Card.cpp