add new SPI function:

  void write(uint8_t data);
  void write16(uint16_t data);
  void write16(uint16_t data, bool msb);
  void writeBytes(uint8_t * data, uint32_t size);
  void writePattern(uint8_t * data, uint8_t size, uint32_t repeat);

Adafruit_ILI9341:
| Benchmark                 | Old (ms) | New (ms) | Speedup     |
|---------------------------|----------|----------|-------------|
| Screen fill               | 1248369  | 278707   |  +347,91%   |
| Text                      | 86102    | 53785    |  +60,09%    |
| Lines                     | 825400   | 536374   |  +53,89%    |
| Horiz/Vert Lines          | 101875   | 24653    |  +313,24%   |
| Rectangles (outline)      | 65720    | 17295    |  +279,99%   |
| Rectangles (filled)       | 2592250  | 579157   |  +347,59%   |
| Circles (filled)          | 411475   | 179454   |  +129,29%   |
| Circles (outline)         | 360002   | 233584   |  +54,12%    |
| Triangles (outline)       | 261772   | 170118   |  +53,88%    |
| Triangles (filled)        | 866951   | 246237   |  +252,08%   |
| Rounded rects (outline)   | 154131   | 81570    |  +88,96%    |
| Rounded rects (filled)    | 2828112  | 660983   |  +327,86%   |
| Total                     | 9802159  | 3061917  |  +220,13%   |