@Imroy 2015-05-12
Part of this appears to be enabling hardware SPI (correct?), so perhaps this should be split up and the commits made more clear i.e Hardware SPI + Adafruit ILI9341 improvements. And you have a few unrelated commits in there as well.