@Links2004 2015-05-12
when i sent the pull request on github there where only on commit.
but times go on an i develop and merge on my branch, so github automatically adds all this later commits automatically. 
If you want to see the steps simply click on one commit.

hardware SPI was working before also but i add functions to get access to more function of the SPI on the ESP8266. like transfer more then one byte.